[sqlalchemy] split all tests in different files using the SQLAlchemyTestMixin
